refactor: dto 명 변경
