+++ backend/src/main/java/com/team5/issue_tracker/issue/parser/IssueSearchRequestParser.java
@@ -1,0 +1,44 @@
+package com.team5.issue_tracker.issue.parser;
+
+import java.util.ArrayList;
+
+import com.team5.issue_tracker.issue.dto.request.IssueSearchRequest;
+
+public class IssueSearchRequestParser {
+  public static IssueSearchRequest fromQueryString(String q) {
+    IssueSearchRequest issueSearchRequest = new IssueSearchRequest();
+    issueSearchRequest.setLabelNames(new ArrayList<>());
+
+    if(q == null || q.isEmpty()) {
+      return issueSearchRequest; // 쿼리 문자열이 비어있으면 기본값 반환
+    }
+
+    String[] parts = q.trim().split("\\s+"); // 공백으로 구분
+
+    for (String part : parts) {
+      String[] pair = part.split(":", 2); // key:value
+      if (pair.length != 2) {
+        continue; // 잘못된 형식은 무시
+      }
+      String key = pair[0].trim();
+      String value = pair[1].trim();
+      switch (key) {
+        case "is" -> {
+          if (!value.equalsIgnoreCase("open") && !value.equalsIgnoreCase("closed")) {
+            continue;
+          }
+          issueSearchRequest.setIsOpen(value.equalsIgnoreCase("open"));
+        }
+        case "assignee" -> issueSearchRequest.setAssigneeName(value); //TODO: 나중에 빌더로 변경하면 편할듯
+        case "label" -> issueSearchRequest.getLabelNames().add(value);
+        case "milestone" -> issueSearchRequest.setMilestoneName(value);
+        case "author" -> issueSearchRequest.setAuthorName(value);
+        default -> {
+          continue;
+        }
+      }
+    }
+
+    return issueSearchRequest;
+  }
+}
